surveytable: Streamlining Complex Survey Estimation and Reliability Assessment in R

Short and understandable commands that generate tabulated, formatted, and rounded survey estimates. Mostly a wrapper for the 'survey' package (Lumley (2004) doi:10.18637/jss.v009.i08 https://CRAN.R-project.org/package=survey) that identifies low-precision estimates using the National Center for Health Statistics (NCHS) presentation standards (Parker et al. (2017) https://www.cdc.gov/nchs/data/series/sr_02/sr02_175.pdf, Parker et al. (2023) doi:10.15620/cdc:124368 ).

See also

Author

Maintainer: Alex Strashny alex.strashny@gmail.com (ORCID)