Tabulate subsets or interactions

Create subsets of the survey using one variable, and tabulate another variable within each of the subsets. Interact two variables and tabulate.

Usage

tab_cross(vr, vrby, max_levels = getOption("surveytable.max_levels"))

tab_subset(
  vr,
  vrby,
  lvls = c(),
  test = FALSE,
  alpha = 0.05,
  p_adjust = FALSE,
  drop_na = getOption("surveytable.drop_na"),
  max_levels = getOption("surveytable.max_levels")
)

Arguments

vr

variable to tabulate

vrby

use this variable to subset the survey

max_levels

a categorical variable can have at most this many levels. Used to avoid printing huge tables.

lvls

(optional) only show these levels of vrby

test

if TRUE, performs a test of association and t-tests for all pairs of levels of vr and vrby. If test is the name of a level of vr, performs a conditional independence test for that level.

alpha

significance level for tests

p_adjust

adjust p-values for multiple comparisons?

drop_na

drop missing values (NA)? Categorical variables only.

Value

A list of tables or a single table.

Details

tab_subset() creates subsets using the levels of vrby, and tabulates vr in each subset. Optionally, only use the lvls levels of vrby. vr can be categorical (factor or character), logical, or numeric.

tab_cross() crosses or interacts vr and vrby and tabulates the new variable. Tables created using tab_subset() and tab_cross() have the same counts but different percentages. With tab_subset(), percentages within each subset add up to 100%. With tab_cross(), percentages across the entire population add up to 100%. Also see var_cross().

test = TRUE performs a test of association between the two variables. Also performs t-tests for all pairs of levels of vr and vrby.

test = "{LEVEL}", where {LEVEL} is a level of vr, performs a conditional independence test to compare the proportion of vr = "{LEVEL}" for different values of vrby.

See also

Examples

set_survey(namcs2019sv)
#>                            Survey info {NAMCS 2019 PUF}                           
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Variables в”Ӯ Observations в”Ӯ Age adjustment в”Ӯ Design                             в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ        33 в”Ӯ        8,250 в”Ӯ None           в”Ӯ Stratified 1 - level Cluster       в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ Sampling design (with replacement) в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ With (398) clusters.               в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ namcs2019sv =                      в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ survey::svydesign(ids = ~CPSUM,    в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ strata = ~CSTRATM, weights =       в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ ~PATWT                             в”Ӯ
#> в”Ӯ           в”Ӯ              в”Ӯ                в”Ӯ , data = namcs2019sv_df)           в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#> 

# For each SEX, tabulate AGER
tab_subset("AGER", "SEX")
#>               Patient age recode (Patient sex = Female) (knowns only) {NAMCS 2019 PUF}              
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level       в”Ӯ     n в”Ӯ      Number в”Ӯ SE (000) в”Ӯ LL (000) в”Ӯ UL (000) в”Ӯ Percent в”Ӯ  SE в”Ӯ   LL в”Ӯ   UL в”Ӯ
#> в”Ӯ             в”Ӯ       в”Ӯ       (000) в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15    в”Ӯ   434 в”Ӯ      59,958 в”Ӯ    7,206 в”Ӯ   47,318 в”Ӯ   75,974 в”Ӯ     9.9 в”Ӯ 1.2 в”Ӯ  7.6 в”Ӯ 12.6 в”Ӯ
#> в”Ӯ years       в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years в”Ӯ   346 в”Ӯ      41,128 в”Ӯ    4,532 в”Ӯ   33,066 в”Ӯ   51,156 в”Ӯ     6.8 в”Ӯ 0.7 в”Ӯ  5.4 в”Ӯ  8.4 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years в”Ӯ   923 в”Ӯ     113,708 в”Ӯ   11,461 в”Ӯ   93,256 в”Ӯ  138,646 в”Ӯ    18.8 в”Ӯ 1.6 в”Ӯ 15.8 в”Ӯ 22.1 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years в”Ӯ 1,253 в”Ӯ     175,978 в”Ӯ   16,009 в”Ӯ  147,153 в”Ӯ  210,450 в”Ӯ    29.1 в”Ӯ 1.7 в”Ӯ 25.7 в”Ӯ 32.6 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74 years в”Ӯ   891 в”Ӯ     120,099 в”Ӯ   11,066 в”Ӯ  100,171 в”Ӯ  143,992 в”Ӯ    19.8 в”Ӯ 1.5 в”Ӯ 17В В  в”Ӯ 22.9 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years    в”Ӯ   762 в”Ӯ      94,173 в”Ӯ   11,085 в”Ӯ   74,682 в”Ӯ  118,751 в”Ӯ    15.6 в”Ӯ 1.5 в”Ӯ 12.8 в”Ӯ 18.7 в”Ӯ
#> в”Ӯ and over    в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#>   N = 4609. Checked NCHS presentation standards. Nothing to report.                                 
#> 
#>                Patient age recode (Patient sex = Male) (knowns only) {NAMCS 2019 PUF}               
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level       в”Ӯ     n в”Ӯ      Number в”Ӯ SE (000) в”Ӯ LL (000) в”Ӯ UL (000) в”Ӯ Percent в”Ӯ  SE в”Ӯ   LL в”Ӯ   UL в”Ӯ
#> в”Ӯ             в”Ӯ       в”Ӯ       (000) в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15    в”Ӯ   453 в”Ӯ      57,959 в”Ӯ    7,728 в”Ӯ   44,570 в”Ӯ   75,371 в”Ӯ    13.4 в”Ӯ 1.7 в”Ӯ 10.3 в”Ӯ 17.1 в”Ӯ
#> в”Ӯ years       в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years в”Ӯ   196 в”Ӯ      23,728 в”Ӯ    4,344 в”Ӯ   16,457 в”Ӯ   34,210 в”Ӯ     5.5 в”Ӯ 0.8 в”Ӯ  4В В  в”Ӯ  7.4 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years в”Ӯ   512 в”Ӯ      56,562 в”Ӯ    7,277 в”Ӯ   43,861 в”Ӯ   72,942 в”Ӯ    13.1 в”Ӯ 1.3 в”Ӯ 10.7 в”Ӯ 15.8 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years в”Ӯ 1,030 в”Ӯ     133,528 в”Ӯ   12,956 в”Ӯ  110,319 в”Ӯ  161,619 в”Ӯ    30.9 в”Ӯ 1.6 в”Ӯ 27.8 в”Ӯ 34.3 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74 years в”Ӯ   770 в”Ӯ      86,766 в”Ӯ    6,767 в”Ӯ   74,409 в”Ӯ  101,176 в”Ӯ    20.1 в”Ӯ 1.5 в”Ӯ 17.3 в”Ӯ 23.1 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years    в”Ӯ   680 в”Ӯ      72,896 в”Ӯ    6,661 в”Ӯ   60,872 в”Ӯ   87,296 в”Ӯ    16.9 в”Ӯ 1.5 в”Ӯ 14В В  в”Ӯ 20.2 в”Ӯ
#> в”Ӯ and over    в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#>   N = 3641. Checked NCHS presentation standards. Nothing to report.                                 
#> 

# Same counts as tab_subset(), but different percentages.
tab_cross("AGER", "SEX")
#>                        (Patient age recode) x (Patient sex) {NAMCS 2019 PUF}                        
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level       в”Ӯ     n в”Ӯ      Number в”Ӯ SE (000) в”Ӯ LL (000) в”Ӯ UL (000) в”Ӯ Percent в”Ӯ  SE в”Ӯ   LL в”Ӯ   UL в”Ӯ
#> в”Ӯ             в”Ӯ       в”Ӯ       (000) в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15    в”Ӯ   434 в”Ӯ      59,958 в”Ӯ    7,206 в”Ӯ   47,318 в”Ӯ   75,974 в”Ӯ     5.8 в”Ӯ 0.7 в”Ӯ  4.5 в”Ӯ  7.3 в”Ӯ
#> в”Ӯ years:      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24       в”Ӯ   346 в”Ӯ      41,128 в”Ӯ    4,532 в”Ӯ   33,066 в”Ӯ   51,156 в”Ӯ     4В В  в”Ӯ 0.4 в”Ӯ  3.2 в”Ӯ  4.9 в”Ӯ
#> в”Ӯ years:      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44       в”Ӯ   923 в”Ӯ     113,708 в”Ӯ   11,461 в”Ӯ   93,256 в”Ӯ  138,646 в”Ӯ    11В В  в”Ӯ 1В В  в”Ӯ  9В В  в”Ӯ 13.2 в”Ӯ
#> в”Ӯ years:      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64       в”Ӯ 1,253 в”Ӯ     175,978 в”Ӯ   16,009 в”Ӯ  147,153 в”Ӯ  210,450 в”Ӯ    17В В  в”Ӯ 1.1 в”Ӯ 14.8 в”Ӯ 19.3 в”Ӯ
#> в”Ӯ years:      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74       в”Ӯ   891 в”Ӯ     120,099 в”Ӯ   11,066 в”Ӯ  100,171 в”Ӯ  143,992 в”Ӯ    11.6 в”Ӯ 1В В  в”Ӯ  9.7 в”Ӯ 13.7 в”Ӯ
#> в”Ӯ years:      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years    в”Ӯ   762 в”Ӯ      94,173 в”Ӯ   11,085 в”Ӯ   74,682 в”Ӯ  118,751 в”Ӯ     9.1 в”Ӯ 0.9 в”Ӯ  7.3 в”Ӯ 11.1 в”Ӯ
#> в”Ӯ and over:   в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Female      в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15    в”Ӯ   453 в”Ӯ      57,959 в”Ӯ    7,728 в”Ӯ   44,570 в”Ӯ   75,371 в”Ӯ     5.6 в”Ӯ 0.7 в”Ӯ  4.3 в”Ӯ  7.2 в”Ӯ
#> в”Ӯ years: Male в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24       в”Ӯ   196 в”Ӯ      23,728 в”Ӯ    4,344 в”Ӯ   16,457 в”Ӯ   34,210 в”Ӯ     2.3 в”Ӯ 0.4 в”Ӯ  1.6 в”Ӯ  3.2 в”Ӯ
#> в”Ӯ years: Male в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44       в”Ӯ   512 в”Ӯ      56,562 в”Ӯ    7,277 в”Ӯ   43,861 в”Ӯ   72,942 в”Ӯ     5.5 в”Ӯ 0.6 в”Ӯ  4.3 в”Ӯ  6.8 в”Ӯ
#> в”Ӯ years: Male в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64       в”Ӯ 1,030 в”Ӯ     133,528 в”Ӯ   12,956 в”Ӯ  110,319 в”Ӯ  161,619 в”Ӯ    12.9 в”Ӯ 1В В  в”Ӯ 10.9 в”Ӯ 15.1 в”Ӯ
#> в”Ӯ years: Male в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74       в”Ӯ   770 в”Ӯ      86,766 в”Ӯ    6,767 в”Ӯ   74,409 в”Ӯ  101,176 в”Ӯ     8.4 в”Ӯ 0.6 в”Ӯ  7.2 в”Ӯ  9.7 в”Ӯ
#> в”Ӯ years: Male в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years    в”Ӯ   680 в”Ӯ      72,896 в”Ӯ    6,661 в”Ӯ   60,872 в”Ӯ   87,296 в”Ӯ     7В В  в”Ӯ 0.6 в”Ӯ  5.9 в”Ӯ  8.3 в”Ӯ
#> в”Ӯ and over:   в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в”Ӯ Male        в”Ӯ       в”Ӯ             в”Ӯ          в”Ӯ          в”Ӯ          в”Ӯ         в”Ӯ     в”Ӯ      в”Ӯ      в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#>   N = 8250. Checked NCHS presentation standards. Nothing to report.                                 
#> 

# Numeric variables
tab_subset("NUMMED", "AGER")
#> Number of medications coded (for different levels of Patient age recode) {NAMCS 2019 PUF}
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level             в”Ӯ % known в”Ӯ Mean в”Ӯ   SEM в”Ӯ   LL в”Ӯ   UL в”Ӯ   SD в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years    в”Ӯ     100 в”Ӯ 1.58 в”Ӯ 0.168 в”Ӯ 1.25 в”Ӯ 1.91 в”Ӯ 1.75 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years       в”Ӯ     100 в”Ӯ 1.64 в”Ӯ 0.112 в”Ӯ 1.42 в”Ӯ 1.86 в”Ӯ 1.7В  в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years       в”Ӯ     100 в”Ӯ 2.15 в”Ӯ 0.225 в”Ӯ 1.71 в”Ӯ 2.59 в”Ӯ 2.74 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years       в”Ӯ     100 в”Ӯ 3.49 в”Ӯ 0.303 в”Ӯ 2.9В  в”Ӯ 4.09 в”Ӯ 4.49 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74 years       в”Ӯ     100 в”Ӯ 4.44 в”Ӯ 0.431 в”Ӯ 3.6В  в”Ӯ 5.29 в”Ӯ 5.03 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years and over в”Ӯ     100 в”Ӯ 5.53 в”Ӯ 0.494 в”Ӯ 4.56 в”Ӯ 6.5В  в”Ӯ 5.59 в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#> 

# Hypothesis testing
tab_subset("NUMMED", "AGER", test = TRUE)
#> Number of medications coded (for different levels of Patient age recode) {NAMCS 2019 PUF}
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level             в”Ӯ % known в”Ӯ Mean в”Ӯ   SEM в”Ӯ   LL в”Ӯ   UL в”Ӯ   SD в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years    в”Ӯ     100 в”Ӯ 1.58 в”Ӯ 0.168 в”Ӯ 1.25 в”Ӯ 1.91 в”Ӯ 1.75 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years       в”Ӯ     100 в”Ӯ 1.64 в”Ӯ 0.112 в”Ӯ 1.42 в”Ӯ 1.86 в”Ӯ 1.7В  в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years       в”Ӯ     100 в”Ӯ 2.15 в”Ӯ 0.225 в”Ӯ 1.71 в”Ӯ 2.59 в”Ӯ 2.74 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years       в”Ӯ     100 в”Ӯ 3.49 в”Ӯ 0.303 в”Ӯ 2.9В  в”Ӯ 4.09 в”Ӯ 4.49 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74 years       в”Ӯ     100 в”Ӯ 4.44 в”Ӯ 0.431 в”Ӯ 3.6В  в”Ӯ 5.29 в”Ӯ 5.03 в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 75 years and over в”Ӯ     100 в”Ӯ 5.53 в”Ӯ 0.494 в”Ӯ 4.56 в”Ӯ 6.5В  в”Ӯ 5.59 в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#> 
#> Association between Number of medications coded and Patient age recode {NAMCS 2019 PUF}
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Test statistic в”Ӯ Degrees of freedom 1 в”Ӯ Degrees of freedom 2 в”Ӯ p-value в”Ӯ Flag в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ             13 в”Ӯ                    5 в”Ӯ                  334 в”Ӯ       0 в”Ӯ *    в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#>   Wald test. *: p <= 0.05                                                        
#> 
#> Comparison of Number of medications coded across all possible pairs of Patient age recode {NAMCS 2019 PUF}
#> в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”җ
#> в”Ӯ Level 1        в”Ӯ Level 2           в”Ӯ Test statistic в”Ӯ Degrees of freedom в”Ӯ p-value в”Ӯ Flag в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years в”Ӯ 15-24 years       в”Ӯ           0.33 в”Ӯ                173 в”Ӯ   0.739 в”Ӯ      в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years в”Ӯ 25-44 years       в”Ӯ           2.03 в”Ӯ                267 в”Ӯ   0.043 в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years в”Ӯ 45-64 years       в”Ӯ           5.59 в”Ӯ                315 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years в”Ӯ 65-74 years       в”Ӯ           6.2В  в”Ӯ                294 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ Under 15 years в”Ӯ 75 years and over в”Ӯ           7.6В  в”Ӯ                269 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years    в”Ӯ 25-44 years       в”Ӯ           2.19 в”Ӯ                268 в”Ӯ   0.029 в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years    в”Ӯ 45-64 years       в”Ӯ           5.84 в”Ӯ                317 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years    в”Ӯ 65-74 years       в”Ӯ           6.35 в”Ӯ                310 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 15-24 years    в”Ӯ 75 years and over в”Ӯ           7.56 в”Ӯ                296 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years    в”Ӯ 45-64 years       в”Ӯ           5.65 в”Ӯ                297 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years    в”Ӯ 65-74 years       в”Ӯ           5.53 в”Ӯ                296 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 25-44 years    в”Ӯ 75 years and over в”Ӯ           7В В В  в”Ӯ                290 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years    в”Ӯ 65-74 years       в”Ӯ           2.71 в”Ӯ                297 в”Ӯ   0.007 в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 45-64 years    в”Ӯ 75 years and over в”Ӯ           5.05 в”Ӯ                291 в”Ӯ   0В В В В  в”Ӯ *    в”Ӯ
#> в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
#> в”Ӯ 65-74 years    в”Ӯ 75 years and over в”Ӯ           3.05 в”Ӯ                280 в”Ӯ   0.002 в”Ӯ *    в”Ӯ
#> в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ
#>   Design-based t-test. *: p <= 0.05                                                          
#>