Convert factor to logical

Convert factor to logical

Usage

var_case(newvr, vr, cases, retain_na = TRUE)

Arguments

newvr

name of the new logical variable to be created

vr

factor variable

cases

one or more levels of vr that are converted to TRUE. All other levels are converted to FALSE.

retain_na

for the observations where vr is NA, should newvr be NA as well?

Value

Survey object

See also

Examples

set_survey(namcs2019sv)
#>                         Survey info {NAMCS 2019 PUF}                         
#> ┌───────────┬──────────────┬────────────────────────────────────────────────┐
#> │ Variables │ Observations │ Design                                         │
#> ├───────────┼──────────────┼────────────────────────────────────────────────┤
#> │        33 │        8,250 │ Stratified 1 - level Cluster Sampling design   │
#> │           │              │ (with replacement)                             │
#> │           │              │ With (398) clusters.                           │
#> │           │              │ namcs2019sv = survey::svydesign(ids = ~ CPSUM, │
#> │           │              │ strata = ~ CSTRATM, weights = ~ PATWT          │
#> │           │              │ , data = namcs2019sv_df)                       │
#> └───────────┴──────────────┴────────────────────────────────────────────────┘
#> 

var_case("Preventive care visits", "MAJOR", "Preventive care")
tab("Preventive care visits")
#>                            Preventive care visits {NAMCS 2019 PUF}                            
#> ┌───────┬───────┬─────────────┬──────────┬──────────┬──────────┬─────────┬─────┬──────┬──────┐
#> │ Level │     n │      Number │ SE (000) │ LL (000) │ UL (000) │ Percent │  SE │   LL │   UL │
#> │       │       │       (000) │          │          │          │         │     │      │      │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ FALSE │ 6,682 │     812,861 │   45,220 │  728,841 │  906,566 │    78.4 │ 1.7 │ 74.9 │ 81.7 │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ TRUE  │ 1,568 │     223,624 │   18,520 │  190,068 │  263,103 │    21.6 │ 1.7 │ 18.3 │ 25.1 │
#> └───────┴───────┴─────────────┴──────────┴──────────┴──────────┴─────────┴─────┴──────┴──────┘
#>   N = 8250. Checked NCHS presentation standards. Nothing to                                   
#>   report.                                                                                     
#> 

var_case("Surgery-related visits"
, "MAJOR"
, c("Pre-surgery", "Post-surgery"))
tab("Surgery-related visits")
#>                             Surgery-related visits {NAMCS 2019 PUF}                            
#> ┌───────┬───────┬─────────────┬──────────┬──────────┬───────────┬─────────┬─────┬──────┬──────┐
#> │ Level │     n │      Number │ SE (000) │ LL (000) │  UL (000) │ Percent │  SE │   LL │   UL │
#> │       │       │       (000) │          │          │           │         │     │      │      │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼───────────┼─────────┼─────┼──────┼──────┤
#> │ FALSE │ 7,432 │     969,451 │   47,976 │  879,793 │ 1,068,246 │    93.5 │ 0.8 │ 91.9 │ 94.9 │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼───────────┼─────────┼─────┼──────┼──────┤
#> │ TRUE  │   818 │      67,034 │    7,810 │   53,273 │    84,348 │     6.5 │ 0.8 │  5.1 │  8.1 │
#> └───────┴───────┴─────────────┴──────────┴──────────┴───────────┴─────────┴─────┴──────┴──────┘
#>   N = 8250. Checked NCHS presentation standards. Nothing to                                    
#>   report.                                                                                      
#> 

var_case("Non-primary"
, "SPECCAT.bad"
, c("Surgical care specialty", "Medical care specialty"))
tab("Non-primary")
#>                                  Non-primary {NAMCS 2019 PUF}                                 
#> ┌───────┬───────┬─────────────┬──────────┬──────────┬──────────┬─────────┬─────┬──────┬──────┐
#> │ Level │     n │      Number │ SE (000) │ LL (000) │ UL (000) │ Percent │  SE │   LL │   UL │
#> │       │       │       (000) │          │          │          │         │     │      │      │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ FALSE │ 2,406 │     422,807 │   26,382 │  374,099 │  477,857 │    40.8 │ 2.2 │ 36.5 │ 45.2 │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ TRUE  │ 4,194 │     406,216 │   32,643 │  346,937 │  475,622 │    39.2 │ 2.1 │ 35   │ 43.5 │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ <N/A> │ 1,650 │     207,462 │   12,458 │  184,378 │  233,436 │    20   │ 0.8 │ 18.5 │ 21.6 │
#> └───────┴───────┴─────────────┴──────────┴──────────┴──────────┴─────────┴─────┴──────┴──────┘
#>   N = 8250. Checked NCHS presentation standards. Nothing to                                   
#>   report.                                                                                     
#> 
tab("Non-primary", drop_na = TRUE)
#>                           Non-primary (knowns only) {NAMCS 2019 PUF}                          
#> ┌───────┬───────┬─────────────┬──────────┬──────────┬──────────┬─────────┬─────┬──────┬──────┐
#> │ Level │     n │      Number │ SE (000) │ LL (000) │ UL (000) │ Percent │  SE │   LL │   UL │
#> │       │       │       (000) │          │          │          │         │     │      │      │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ FALSE │ 2,406 │     422,807 │   26,382 │  374,099 │  477,857 │      51 │ 2.6 │ 45.7 │ 56.3 │
#> ├───────┼───────┼─────────────┼──────────┼──────────┼──────────┼─────────┼─────┼──────┼──────┤
#> │ TRUE  │ 4,194 │     406,216 │   32,643 │  346,937 │  475,622 │      49 │ 2.6 │ 43.7 │ 54.3 │
#> └───────┴───────┴─────────────┴──────────┴──────────┴──────────┴─────────┴─────┴──────┴──────┘
#>   N = 6600. Checked NCHS presentation standards. Nothing to                                   
#>   report.                                                                                     
#>