Tabulate subsets or interactions

Create subsets of the survey using one variable, and tabulate another variable within each of the subsets. Interact two variables and tabulate. Test equality of proportions or means.

Usage

tab_cross(
  vr,
  vrby,
  max_levels = getOption("surveytable.max_levels"),
  screen = getOption("surveytable.screen"),
  csv = getOption("surveytable.csv")
)

tab_subset(
  vr,
  vrby,
  lvls = c(),
  test = FALSE,
  alpha = 0.05,
  drop_na = getOption("surveytable.drop_na"),
  max_levels = getOption("surveytable.max_levels"),
  screen = getOption("surveytable.screen"),
  csv = getOption("surveytable.csv")
)

Arguments

vr

variable to tabulate

vrby

use this variable to subset the survey

max_levels

a categorical variable can have at most this many levels. Used to avoid printing huge tables.

screen

print to the screen?

csv

name of a CSV file

lvls

(optional) only show these levels of vrby

test

perform hypothesis test?

alpha

significance level for the above test.

drop_na

drop missing values (NA)? Categorical variables only.

Value

tab_subset: A list of data.frame tables or a single data.frame table.

tab_cross: A data.frame table.

Details

tab_subset creates subsets using the levels of vrby, and tabulates vr in each subset. Optionally, only use the lvls levels of vrby. vr can be categorical (factor), logical, or numeric.

tab_cross crosses or interacts vr and vrby and tabulates the new variable. Tables created using tab_subset and tab_cross have the same counts but different percentages. With tab_subset, percentages within each subset add up to 100%. With tab_cross, percentages across the entire population add up to 100%. Also see var_cross().

See also

Examples

set_survey("namcs2019sv")
#>                        _             
#> Survey name            NAMCS 2019 PUF
#> Number of variables    38            
#> Number of observations 8250          
#> Stratified 1 - level Cluster Sampling design (with replacement)
#> With (398) clusters.
#> svydesign(ids = ~CPSUM, strata = ~CSTRATM, weights = ~PATWT, 
#>     data = namcs2019sv_df)
#> * To adjust how counts are rounded, see ?set_count_int

# For each SEX, tabulate AGER
tab_subset("AGER", "SEX")
#>                   Patient age recode (Patient sex = Female) {NAMCS 2019 PUF}                   
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       59,958 ¦    7,206 ¦   47,318 ¦   75,974 ¦     9.9 ¦ 1.2 ¦  7.6 ¦ 12.6 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       41,128 ¦    4,532 ¦   33,066 ¦   51,156 ¦     6.8 ¦ 0.7 ¦  5.4 ¦  8.4 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years   ¦      113,708 ¦   11,461 ¦   93,256 ¦  138,646 ¦    18.8 ¦ 1.6 ¦ 15.8 ¦ 22.1 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years   ¦      175,978 ¦   16,009 ¦  147,153 ¦  210,450 ¦    29.1 ¦ 1.7 ¦ 25.7 ¦ 32.6 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years   ¦      120,099 ¦   11,066 ¦  100,171 ¦  143,992 ¦    19.8 ¦ 1.5 ¦ 17   ¦ 22.9 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       94,173 ¦   11,085 ¦   74,682 ¦  118,751 ¦    15.6 ¦ 1.5 ¦ 12.8 ¦ 18.7 ¦
#> ¦ over          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 
#>                    Patient age recode (Patient sex = Male) {NAMCS 2019 PUF}                    
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       57,959 ¦    7,728 ¦   44,570 ¦   75,371 ¦    13.4 ¦ 1.7 ¦ 10.3 ¦ 17.1 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       23,728 ¦    4,344 ¦   16,457 ¦   34,210 ¦     5.5 ¦ 0.8 ¦  4   ¦  7.4 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years   ¦       56,562 ¦    7,277 ¦   43,861 ¦   72,942 ¦    13.1 ¦ 1.3 ¦ 10.7 ¦ 15.8 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years   ¦      133,528 ¦   12,956 ¦  110,319 ¦  161,619 ¦    30.9 ¦ 1.6 ¦ 27.8 ¦ 34.3 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years   ¦       86,766 ¦    6,767 ¦   74,409 ¦  101,176 ¦    20.1 ¦ 1.5 ¦ 17.3 ¦ 23.1 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       72,896 ¦    6,661 ¦   60,872 ¦   87,296 ¦    16.9 ¦ 1.5 ¦ 14   ¦ 20.2 ¦
#> ¦ over          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 

# Same counts as tab_subset(), but different percentages.
tab_cross("AGER", "SEX")
#>                      (Patient age recode) x (Patient sex) {NAMCS 2019 PUF}                     
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       59,958 ¦    7,206 ¦   47,318 ¦   75,974 ¦     5.8 ¦ 0.7 ¦  4.5 ¦  7.3 ¦
#> ¦ years :       ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> ¦ Female        ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years : ¦       41,128 ¦    4,532 ¦   33,066 ¦   51,156 ¦     4   ¦ 0.4 ¦  3.2 ¦  4.9 ¦
#> ¦ Female        ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years : ¦      113,708 ¦   11,461 ¦   93,256 ¦  138,646 ¦    11   ¦ 1   ¦  9   ¦ 13.2 ¦
#> ¦ Female        ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years : ¦      175,978 ¦   16,009 ¦  147,153 ¦  210,450 ¦    17   ¦ 1.1 ¦ 14.8 ¦ 19.3 ¦
#> ¦ Female        ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years : ¦      120,099 ¦   11,066 ¦  100,171 ¦  143,992 ¦    11.6 ¦ 1   ¦  9.7 ¦ 13.7 ¦
#> ¦ Female        ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       94,173 ¦   11,085 ¦   74,682 ¦  118,751 ¦     9.1 ¦ 0.9 ¦  7.3 ¦ 11.1 ¦
#> ¦ over : Female ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       57,959 ¦    7,728 ¦   44,570 ¦   75,371 ¦     5.6 ¦ 0.7 ¦  4.3 ¦  7.2 ¦
#> ¦ years : Male  ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years : ¦       23,728 ¦    4,344 ¦   16,457 ¦   34,210 ¦     2.3 ¦ 0.4 ¦  1.6 ¦  3.2 ¦
#> ¦ Male          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years : ¦       56,562 ¦    7,277 ¦   43,861 ¦   72,942 ¦     5.5 ¦ 0.6 ¦  4.3 ¦  6.8 ¦
#> ¦ Male          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years : ¦      133,528 ¦   12,956 ¦  110,319 ¦  161,619 ¦    12.9 ¦ 1   ¦ 10.9 ¦ 15.1 ¦
#> ¦ Male          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years : ¦       86,766 ¦    6,767 ¦   74,409 ¦  101,176 ¦     8.4 ¦ 0.6 ¦  7.2 ¦  9.7 ¦
#> ¦ Male          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       72,896 ¦    6,661 ¦   60,872 ¦   87,296 ¦     7   ¦ 0.6 ¦  5.9 ¦  8.3 ¦
#> ¦ over : Male   ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 

# Hypothesis testing
tab_subset("AGER", "SEX", test = TRUE)
#>                   Patient age recode (Patient sex = Female) {NAMCS 2019 PUF}                   
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       59,958 ¦    7,206 ¦   47,318 ¦   75,974 ¦     9.9 ¦ 1.2 ¦  7.6 ¦ 12.6 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       41,128 ¦    4,532 ¦   33,066 ¦   51,156 ¦     6.8 ¦ 0.7 ¦  5.4 ¦  8.4 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years   ¦      113,708 ¦   11,461 ¦   93,256 ¦  138,646 ¦    18.8 ¦ 1.6 ¦ 15.8 ¦ 22.1 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years   ¦      175,978 ¦   16,009 ¦  147,153 ¦  210,450 ¦    29.1 ¦ 1.7 ¦ 25.7 ¦ 32.6 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years   ¦      120,099 ¦   11,066 ¦  100,171 ¦  143,992 ¦    19.8 ¦ 1.5 ¦ 17   ¦ 22.9 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       94,173 ¦   11,085 ¦   74,682 ¦  118,751 ¦    15.6 ¦ 1.5 ¦ 12.8 ¦ 18.7 ¦
#> ¦ over          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 
#> Comparison of all possible pairs of Patient age recode (Patient sex = Female) {NAMCS 2019 PUF}
#> +-----------------------------------------------------+
#> ¦ Level 1        ¦ Level 2           ¦ p-value ¦ Flag ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 15-24 years       ¦   0.01  ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 25-44 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 65-74 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 75 years and over ¦   0.007 ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 25-44 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 65-74 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 75 years and over ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 65-74 years       ¦   0.677 ¦      ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 75 years and over ¦   0.201 ¦      ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 45-64 years    ¦ 65-74 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 45-64 years    ¦ 75 years and over ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 65-74 years    ¦ 75 years and over ¦   0.034 ¦ *    ¦
#> +-----------------------------------------------------+
#>   *: p-value <= 0.05                                   
#> 
#>                    Patient age recode (Patient sex = Male) {NAMCS 2019 PUF}                    
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦       57,959 ¦    7,728 ¦   44,570 ¦   75,371 ¦    13.4 ¦ 1.7 ¦ 10.3 ¦ 17.1 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       23,728 ¦    4,344 ¦   16,457 ¦   34,210 ¦     5.5 ¦ 0.8 ¦  4   ¦  7.4 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years   ¦       56,562 ¦    7,277 ¦   43,861 ¦   72,942 ¦    13.1 ¦ 1.3 ¦ 10.7 ¦ 15.8 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years   ¦      133,528 ¦   12,956 ¦  110,319 ¦  161,619 ¦    30.9 ¦ 1.6 ¦ 27.8 ¦ 34.3 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years   ¦       86,766 ¦    6,767 ¦   74,409 ¦  101,176 ¦    20.1 ¦ 1.5 ¦ 17.3 ¦ 23.1 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦       72,896 ¦    6,661 ¦   60,872 ¦   87,296 ¦    16.9 ¦ 1.5 ¦ 14   ¦ 20.2 ¦
#> ¦ over          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 
#> Comparison of all possible pairs of Patient age recode (Patient sex = Male) {NAMCS 2019 PUF}
#> +-----------------------------------------------------+
#> ¦ Level 1        ¦ Level 2           ¦ p-value ¦ Flag ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 15-24 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 25-44 years       ¦   0.894 ¦      ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 65-74 years       ¦   0.008 ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ Under 15 years ¦ 75 years and over ¦   0.166 ¦      ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 25-44 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 65-74 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 15-24 years    ¦ 75 years and over ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 45-64 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 65-74 years       ¦   0.001 ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 25-44 years    ¦ 75 years and over ¦   0.113 ¦      ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 45-64 years    ¦ 65-74 years       ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 45-64 years    ¦ 75 years and over ¦   0     ¦ *    ¦
#> +----------------+-------------------+---------+------¦
#> ¦ 65-74 years    ¦ 75 years and over ¦   0.105 ¦      ¦
#> +-----------------------------------------------------+
#>   *: p-value <= 0.05                                   
#> 

# Numeric variables
tab_subset("NUMMED", "AGER")
#> Number of medications coded (for different levels of Patient age recode) {NAMCS 2019 PUF}
#> +---------------------------------------------------+
#> ¦ Level             ¦ % known ¦ Mean ¦   SEM ¦   SD ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ Under 15 years    ¦     100 ¦ 1.58 ¦ 0.168 ¦ 1.75 ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ 15-24 years       ¦     100 ¦ 1.64 ¦ 0.112 ¦ 1.7  ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ 25-44 years       ¦     100 ¦ 2.15 ¦ 0.225 ¦ 2.74 ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ 45-64 years       ¦     100 ¦ 3.49 ¦ 0.303 ¦ 4.49 ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ 65-74 years       ¦     100 ¦ 4.44 ¦ 0.431 ¦ 5.03 ¦
#> +-------------------+---------+------+-------+------¦
#> ¦ 75 years and over ¦     100 ¦ 5.53 ¦ 0.494 ¦ 5.59 ¦
#> +---------------------------------------------------+
#>