Is any variable true? (Logical OR)

Create a new variable which is true if any of the variables in a list of variables are true.

Usage

var_any(newvr, vrs)

Arguments

newvr

name of the new variable to be created

vrs

vector of logical variables

Value

Survey object

See also

Examples

set_survey(namcs2019sv)
#>                         Survey info {NAMCS 2019 PUF}                         
#> ┌───────────┬──────────────┬────────────────────────────────────────────────┐
#> │ Variables │ Observations │ Design                                         │
#> ├───────────┼──────────────┼────────────────────────────────────────────────┤
#> │        33 │        8,250 │ Stratified 1 - level Cluster Sampling design   │
#> │           │              │ (with replacement)                             │
#> │           │              │ With (398) clusters.                           │
#> │           │              │ namcs2019sv = survey::svydesign(ids = ~CPSUM,  │
#> │           │              │ strata = ~CSTRATM, weights = ~PATWT            │
#> │           │              │ , data = namcs2019sv_df)                       │
#> └───────────┴──────────────┴────────────────────────────────────────────────┘
#> 
var_any("Imaging services"
, c("ANYIMAGE", "BONEDENS", "CATSCAN", "ECHOCARD", "OTHULTRA"
, "MAMMO", "MRI", "XRAY", "OTHIMAGE"))
tab("Imaging services")
#>                                    Imaging services {NAMCS 2019 PUF}                                   
#> ┌───────┬───────┬─────────────┬────────────┬─────────────┬─────────────┬─────────┬──────┬──────┬──────┐
#> │ Level │     n │      Number │         SE │          LL │          UL │ Percent │ SE.1 │ LL.1 │ UL.1 │
#> ├───────┼───────┼─────────────┼────────────┼─────────────┼─────────────┼─────────┼──────┼──────┼──────┤
#> │ FALSE │ 7,148 │ 901,115,076 │ 43,298,146 │ 820,085,161 │ 990,151,291 │    86.9 │  1.1 │ 84.6 │ 89.1 │
#> ├───────┼───────┼─────────────┼────────────┼─────────────┼─────────────┼─────────┼──────┼──────┼──────┤
#> │ TRUE  │ 1,102 │ 135,369,280 │ 13,573,736 │ 111,133,847 │ 164,889,838 │    13.1 │  1.1 │ 10.9 │ 15.4 │
#> └───────┴───────┴─────────────┴────────────┴─────────────┴─────────────┴─────────┴──────┴──────┴──────┘
#>   N = 8250.                                                                                            
#>