Copy a variable

Create a new variable that is a copy of another variable. You can modify the copy, while the original remains unchanged. See examples.

Usage

var_copy(newvr, vr)

Arguments

newvr

name of the new variable to be created

vr

variable

Value

Survey object

See also

Examples

set_survey(namcs2019sv)
#> * To adjust how counts are rounded, see ?set_count_int
#>                                      Survey info {NAMCS 2019 PUF}                                     
#> +----------------------------------------------------------------------------------------------------+
#> ¦ Survey name    ¦ Number of variables ¦ Number of observations ¦ Design                             ¦
#> +----------------+---------------------+------------------------+------------------------------------¦
#> ¦ NAMCS 2019 PUF ¦                  33 ¦                  8,250 ¦ Stratified 1 - level Cluster       ¦
#> ¦                ¦                     ¦                        ¦ Sampling design (with replacement) ¦
#> ¦                ¦                     ¦                        ¦ With (398) clusters.               ¦
#> ¦                ¦                     ¦                        ¦ survey::svydesign(ids = ~CPSUM,    ¦
#> ¦                ¦                     ¦                        ¦ strata = ~CSTRATM, weights =       ¦
#> ¦                ¦                     ¦                        ¦ ~PATWT,                            ¦
#> ¦                ¦                     ¦                        ¦ data = namcs2019sv_df)             ¦
#> +----------------------------------------------------------------------------------------------------+
#> 
var_copy("Age group", "AGER")
var_collapse("Age group", "65+", c("65-74 years", "75 years and over"))
var_collapse("Age group", "25-64", c("25-44 years", "45-64 years"))
tab("AGER", "Age group")
#>                               Patient age recode {NAMCS 2019 PUF}                              
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦      117,917 ¦   14,097 ¦   93,229 ¦  149,142 ¦    11.4 ¦ 1.3 ¦  8.9 ¦ 14.2 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       64,856 ¦    7,018 ¦   52,387 ¦   80,292 ¦     6.3 ¦ 0.6 ¦  5.1 ¦  7.5 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-44 years   ¦      170,271 ¦   13,966 ¦  144,925 ¦  200,049 ¦    16.4 ¦ 1.1 ¦ 14.3 ¦ 18.8 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 45-64 years   ¦      309,506 ¦   23,290 ¦  266,994 ¦  358,787 ¦    29.9 ¦ 1.4 ¦ 27.2 ¦ 32.6 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65-74 years   ¦      206,866 ¦   14,366 ¦  180,481 ¦  237,109 ¦    20   ¦ 1.2 ¦ 17.6 ¦ 22.5 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 75 years and  ¦      167,069 ¦   15,179 ¦  139,746 ¦  199,735 ¦    16.1 ¦ 1.3 ¦ 13.7 ¦ 18.8 ¦
#> ¦ over          ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#> 
#>                                   Age group {NAMCS 2019 PUF}                                   
#> +---------------------------------------------------------------------------------------------+
#> ¦ Level         ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 15      ¦      117,917 ¦   14,097 ¦   93,229 ¦  149,142 ¦    11.4 ¦ 1.3 ¦  8.9 ¦ 14.2 ¦
#> ¦ years         ¦              ¦          ¦          ¦          ¦         ¦     ¦      ¦      ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-24 years   ¦       64,856 ¦    7,018 ¦   52,387 ¦   80,292 ¦     6.3 ¦ 0.6 ¦  5.1 ¦  7.5 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 25-64         ¦      479,777 ¦   32,175 ¦  420,624 ¦  547,247 ¦    46.3 ¦ 1.8 ¦ 42.7 ¦ 49.9 ¦
#> +---------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65+           ¦      373,935 ¦   24,523 ¦  328,777 ¦  425,296 ¦    36.1 ¦ 1.9 ¦ 32.3 ¦ 40   ¦
#> +---------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                         
#>