Convert numeric to factor

Create a new categorical variable based on a numeric variable.

Usage

var_cut(newvr, vr, breaks, labels)

Arguments

newvr

name of the new factor variable to be created

vr

numeric variable

breaks

see cut()

labels

see cut()

Value

(Nothing.)

See also

Examples

set_survey("namcs2019sv")
#>                        _             
#> Survey name            NAMCS 2019 PUF
#> Number of variables    38            
#> Number of observations 8250          
#> Stratified 1 - level Cluster Sampling design (with replacement)
#> With (398) clusters.
#> survey::svydesign(ids = ~CPSUM, strata = ~CSTRATM, weights = ~PATWT, 
#>     data = namcs2019sv_df)
#> * To adjust how counts are rounded, see ?set_count_int
var_cut("Age group"
, "AGE"
, c(-Inf, 0, 4, 14, 64, Inf)
, c("Under 1", "1-4", "5-14", "15-64", "65 and over"))
#> Warning: Age group: overwriting a variable that already exists.
tab("Age group")
#>                                  Age group {NAMCS 2019 PUF}                                  
#> +-------------------------------------------------------------------------------------------+
#> ¦ Level       ¦ Number (000) ¦ SE (000) ¦ LL (000) ¦ UL (000) ¦ Percent ¦  SE ¦   LL ¦   UL ¦
#> +-------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ Under 1     ¦       31,148 ¦    5,282 ¦   22,269 ¦   43,566 ¦     3   ¦ 0.5 ¦  2.1 ¦  4.1 ¦
#> +-------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 1-4         ¦       38,240 ¦    5,444 ¦   28,864 ¦   50,662 ¦     3.7 ¦ 0.5 ¦  2.7 ¦  4.8 ¦
#> +-------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 5-14        ¦       48,529 ¦    5,741 ¦   38,430 ¦   61,282 ¦     4.7 ¦ 0.5 ¦  3.7 ¦  5.9 ¦
#> +-------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 15-64       ¦      544,632 ¦   36,082 ¦  478,254 ¦  620,223 ¦    52.5 ¦ 2   ¦ 48.6 ¦ 56.5 ¦
#> +-------------+--------------+----------+----------+----------+---------+-----+------+------¦
#> ¦ 65 and over ¦      373,935 ¦   24,523 ¦  328,777 ¦  425,296 ¦    36.1 ¦ 1.9 ¦ 32.3 ¦ 40   ¦
#> +-------------------------------------------------------------------------------------------+
#>   (Checked presentation standards. Nothing to report.)                                       
#>